Prabhashi

+++ src/app/programs/delete-program/delete-program.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import {ActivatedRoute, Router} from '@angular/router';
 import { ProgramService } from 'src/app/services/program.service';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-delete-program',
@@ -15,26 +15,73 @@ export class DeleteProgramComponent implements OnInit {
   ProgramId: String= '';
   constructor(private activatedRoute: ActivatedRoute,
      private ProgramService : ProgramService,
-    private _snackBar: MatSnackBar,
     private router: Router) { }
 
   ngOnInit(): void {
     this.activatedRoute.params.subscribe(data=>{
       this.ProgramId=data.programId;
-    });
-
-    if(this.ProgramId){
+    })
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: true,
+    confirmButtonColor: 'limegreen',
+    cancelButtonColor:'red'
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, Delete!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed){
+       
+        if(this.ProgramId){
       
-      this.ProgramService.deleteProgram(this.ProgramId).subscribe(data=>{
-        this._snackBar.open("Program deleted Successfully");
-        this.router.navigateByUrl('programs/list');
-      },err=>{
-        this._snackBar.open("Unable to delete the Program");
-
-      }
-        )
+          this.ProgramService.deleteProgram(this.ProgramId).subscribe(data=>{
+            swalWithBootstrapButtons.fire({
+              position: 'center',
+              icon: 'success',
+              title: 'Deleted!',
+              text:'Event Deleted Successfully!',
+              showConfirmButton: false,
+              timerProgressBar: true,
+              timer: 2500,
+              buttonsStyling: true,
+              confirmButtonColor: 'limegreen', 
+            })
+          },err=>{
+            swalWithBootstrapButtons.fire({
+              position: 'center',
+              icon: 'error',
+              title: 'Unable to Delete Event!',
+              showConfirmButton: false,
+              timerProgressBar: true,
+              timer: 2500
+            })
+          })
+            
+        }
+          
+        } else if (
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire(
+            'Cancelled',
+            '',
+            'error'
+          )
+        }
+        this.router.navigateByUrl('/programs/list');
+      })
+      
     }
-  }
-
-}
-
+    
+    }
+
